+++ spec/api/logistic/order/externalDeliveryBox.spec.js
@@ -7,6 +7,165 @@ const warehouses = require('../../../../warehouses');
 const moment = require('moment');
 const utils = require('../utils');
 
+describe('POST Search ScanExternalDeliveryBox', () => {
+
+  let palladiumClerk = {
+    aid: null,
+    jar: null
+  };
+
+  let hubClerk = {
+    aid: null,
+    jar: null
+  };
+
+  let products, hubWarehouse, palladiumWarehouse;
+  beforeEach(async (done) => {
+    try {
+
+      await lib.dbHelpers.dropAll();
+
+      let warehouse = await models()['WarehouseTest'].insertMany(warehouses);
+      warehouse = JSON.parse(JSON.stringify(warehouse));
+
+      hubWarehouse = warehouse.find(x => x.is_hub && !x.has_customer_pickup);
+
+      palladiumWarehouse = warehouse.find(x => !x.is_hub && x.priority === 1);
+
+      let res1 = await lib.dbHelpers.addAndLoginAgent('hubclerk', _const.ACCESS_LEVEL.HubClerk, hubWarehouse._id);
+      hubClerk.aid = res1.aid;
+      hubClerk.jar = res1.rpJar;
+
+      let res2 = await lib.dbHelpers.addAndLoginAgent('pclerk', _const.ACCESS_LEVEL.ShopClerk, palladiumWarehouse._id);
+      palladiumClerk.aid = res2.aid;
+      palladiumClerk.jar = res2.rpJar;
+
+
+      products = await models()['ProductTest'].insertMany([
+        {
+          name: 'sample 1',
+          base_price: 30000,
+          article_no: "ar123",
+          instances: [
+            {
+              size: "9",
+              price: 2000,
+              barcode: '0394081341',
+            },
+            {
+              size: "10",
+              price: 2000,
+              barcode: '0394081342',
+            }
+          ]
+        }
+      ]);
+
+      done()
+    }
+    catch (err) {
+      console.log(err);
+    }
+  }, 15000);
+
+  it('should get cc order lines which is in its destination', async function (done) {
+    try {
+      this.done = done;
+
+      let orders = [];
+
+      orders.push({
+        customer_id: mongoose.Types.ObjectId(),
+        is_cart: false,
+        transaction_id: "xyz12213",
+        order_lines: [],
+        is_collect: true,
+        order_time: moment(),
+        address: {
+          _id: mongoose.Types.ObjectId(),
+          city: "تهران",
+          street: "مقدس اردبیلی",
+          province: "تهران",
+          warehouse_name: "پالادیوم",
+          warehouse_id: palladiumWarehouse._id,
+          recipient_first_name: "s",
+          recipient_surname: "v",
+          recipient_national_id: "8798789798",
+          recipient_mobile_no: "09124077685",
+        },
+        tickets: [
+          {
+            is_processed: false,
+            status: _const.ORDER_STATUS.ReadyToDeliver,
+            desc: null,
+            receiver_id: palladiumWarehouse._id,
+            timestamp: moment()
+          }
+        ]
+      });
+
+      for (let i = 0; i < 3; i++) { // add 3 order line of first product for order 1
+        orders[0].order_lines.push({
+          paid_price: 0,
+          product_id: products[0].id,
+          product_instance_id: products[0].instances[0].id,
+          adding_time: moment(),
+          tickets: [
+            {
+              is_processed: false,
+              status: _const.ORDER_LINE_STATUS.FinalCheck,
+              desc: null,
+              receiver_id: palladiumWarehouse._id,
+              timestamp: moment()
+            }
+          ]
+        })
+      }
+      for (let i = 0; i < 2; i++) { // add 2 order line of second product instance for order 1
+        orders[0].order_lines.push({
+          paid_price: 0,
+          product_id: products[0].id,
+          product_instance_id: products[0].instances[1].id,
+          adding_time: moment(),
+          tickets: [
+            {
+              is_processed: false,
+              status: _const.ORDER_LINE_STATUS.Recieved,
+              desc: null,
+              receiver_id: palladiumWarehouse._id,
+              timestamp: moment()
+            }
+          ]
+        })
+      }
+
+      orders = await models()['OrderTest'].insertMany(orders);
+      orders = JSON.parse(JSON.stringify(orders));
+
+      let res = await rp({
+        method: 'post',
+        uri: lib.helpers.apiTestURL(`search/Ticket`),
+        body: {
+          options: {
+            type: 'ScanToCustomerDelivery',
+          },
+          offset: 0,
+          limit: 10,
+        },
+        json: true,
+        jar: palladiumClerk.jar,
+        resolveWithFullResponse: true
+      });
+      expect(res.statusCode).toBe(200);
+      expect(res.body.data.length).toBe(1);
+      expect(res.body.data[0].total_order_lines).toBe(5);
+      done();
+    } catch (err) {
+      lib.helpers.errorHandler.bind(this)(err);
+    }
+  });
+
+});
 
 describe('External Unassigned Delivery in App', () => {
 
@@ -24,7 +183,7 @@ describe('External Unassigned Delivery in App', () => {
   beforeEach(async done => {
     try {
       await lib.dbHelpers.dropAll();
-      const agent = await lib.dbHelpers.addAndLoginAgent('IDelivery Agent', _const.ACCESS_LEVEL.InternalDeliveryAgent)
+      const agent = await lib.dbHelpers.addAndLoginAgent('Delivery Agent', _const.ACCESS_LEVEL.DeliveryAgent)
       agentObj.aid = agent.aid;
       agentObj.jar = agent.rpJar;
       await models()['WarehouseTest'].insertMany(warehouses);
@@ -49,7 +208,13 @@ describe('External Unassigned Delivery in App', () => {
               discount_ref: 0
             },
             product_instance_id: mongoose.Types.ObjectId(products[0].instances[0]._id),
-            tickets: []
+            tickets: [{
+              is_processed: false,
+              status: _const.ORDER_LINE_STATUS.Delivered,
+              desc: null,
+              receiver_id: warehouses.find(x => x.is_hub)._id,
+              timestamp: new Date()
+            }]
           },
           tickets: [{
             is_processed: false,
@@ -71,13 +236,19 @@ describe('External Unassigned Delivery in App', () => {
               discount_ref: 0
             },
             product_instance_id: mongoose.Types.ObjectId(products[0].instances[0]._id),
-            tickets: []
+            tickets: [{
+              is_processed: false,
+              status: _const.ORDER_LINE_STATUS.ReturnRequested,
+              desc: null,
+              receiver_id: customerobj._id,
+              timestamp: new Date()
+            }]
           },
           tickets: [{
             is_processed: false,
-            status: _const.ORDER_STATUS.DeliverySet,
+            status: _const.ORDER_STATUS.Delivered,
             desc: null,
-            receiver_id: mongoose.Types.ObjectId(),
+            receiver_id: customerobj._id,
             timestamp: new Date()
           }]
         }
@@ -100,9 +271,12 @@ describe('External Unassigned Delivery in App', () => {
             ],
             _id: mongoose.Types.ObjectId(),
             order_id: orders[0]._id
-
           }],
           start: new Date(),
+          slot: [{
+            lower_bound: 18,
+            upper_bound: 22
+          }],
           tickets: [{
             is_processed: false,
             _id: mongoose.Types.ObjectId(),
@@ -110,6 +284,7 @@ describe('External Unassigned Delivery in App', () => {
             receiver_id: agentObj.aid,
             timestamp: new Date()
           }],
+          shelf_code: "A",
         },
         {
           to: {
@@ -133,12 +308,11 @@ describe('External Unassigned Delivery in App', () => {
           tickets: [{
             is_processed: false,
             _id: mongoose.Types.ObjectId(),
-            status: _const.DELIVERY_STATUS.default,
-            receiver_id: mongoose.Types.ObjectId(),
+            status: _const.DELIVERY_STATUS.agentSet,
+            receiver_id: agentObj._id,
             timestamp: new Date()
           }],
-
-        },
+        }
       ]);
       deliveries = JSON.parse(JSON.stringify(deliveries));
       done();
@@ -149,7 +323,7 @@ describe('External Unassigned Delivery in App', () => {
   }, 15000);
 
 
-  it('should see deliveries from hub to customer and customer to hub', async function (done) {
+  it('should see delivery from hub to customer', async function (done) {
     this.done = done;
 
     const res = await rp({
@@ -168,7 +342,30 @@ describe('External Unassigned Delivery in App', () => {
     });
 
     expect(res.statusCode).toBe(200);
-    expect(res.body.data.length).toBe(2);
+    expect(res.body.data.length).toBe(1);
+    done();
+  });
+
+  it('should see delivery return from customer to hub', async function (done) {
+    this.done = done;
+
+    const res = await rp({
+      uri: lib.helpers.apiTestURL('search/DeliveryTicket'),
+      method: 'POST',
+      body: {
+        options: {
+          type: 'ExternalUnassignedDelivery'
+        },
+        offset: 0,
+        limit: 10
+      },
+      json: true,
+      jar: agentObj.jar,
+      resolveWithFullResponse: true
+    });
+
+    expect(res.statusCode).toBe(200);
+    expect(res.body.data.length).toBe(1);
     done();
   });
 
@@ -190,7 +387,7 @@ describe('On Delivery in App', () => {
   beforeEach(async done => {
     try {
       await lib.dbHelpers.dropAll();
-      const agent = await lib.dbHelpers.addAndLoginAgent('IDelivery Agent', _const.ACCESS_LEVEL.InternalDeliveryAgent)
+      const agent = await lib.dbHelpers.addAndLoginAgent('Delivery Agent', _const.ACCESS_LEVEL.DeliveryAgent)
       agentObj.aid = agent.aid;
       agentObj.jar = agent.rpJar;
       await models()['WarehouseTest'].insertMany(warehouses);
@@ -204,7 +401,7 @@ describe('On Delivery in App', () => {
 
       products = await utils.makeProducts();
       orders = await utils.makeOrders(customer);
-      await models()['OrderTest'].update({
+      await models()['OrderTest'].updateMany({
         _id: mongoose.Types.ObjectId(orders[0]._id),
       }, {
         $set: {
@@ -215,33 +412,17 @@ describe('On Delivery in App', () => {
               discount_ref: 0
             },
             product_instance_id: mongoose.Types.ObjectId(products[0].instances[0]._id),
-            tickets: []
+            tickets: [{
+              is_processed: false,
+              status: _const.ORDER_LINE_STATUS.Delivered,
+              desc: null,
+              receiver_id: warehouses.find(x => x.is_hub)._id,
+              timestamp: new Date()
+            }]
           },
           tickets: [{
             is_processed: false,
-            status: _const.ORDER_STATUS.DeliverySet,
-            desc: null,
-            receiver_id: mongoose.Types.ObjectId(),
-            timestamp: new Date()
-          }]
-        }
-      });
-      await models()['OrderTest'].update({
-        _id: mongoose.Types.ObjectId(orders[1]._id),
-      }, {
-        $set: {
-          order_lines: {
-            product_id: mongoose.Types.ObjectId(products[0]._id),
-            campaign_info: {
-              _id: mongoose.Types.ObjectId(),
-              discount_ref: 0
-            },
-            product_instance_id: mongoose.Types.ObjectId(products[0].instances[0]._id),
-            tickets: []
-          },
-          tickets: [{
-            is_processed: false,
-            status: _const.ORDER_STATUS.DeliverySet,
+            status: _const.ORDER_STATUS.Delivered,
             desc: null,
             receiver_id: mongoose.Types.ObjectId(),
             timestamp: new Date()
@@ -252,13 +433,13 @@ describe('On Delivery in App', () => {
       deliveries = await models()['DeliveryTest'].insertMany([
         {
           to: {
+            warehouse_id: warehouses.find(x => x.is_hub)._id
+          },
+          from: {
             customer: {
               _id: orderData[0].customer_id,
               address: orderData[0].address
             }
-          },
-          from: {
-            warehouse_id: warehouses.find(x => x.is_hub)._id
           },
           order_details: [{
             order_line_ids: [
@@ -331,7 +512,7 @@ describe('AgentFinishedDelivery in App', () => {
   beforeEach(async done => {
     try {
       await lib.dbHelpers.dropAll();
-      const agent = await lib.dbHelpers.addAndLoginAgent('IDelivery Agent', _const.ACCESS_LEVEL.InternalDeliveryAgent)
+      const agent = await lib.dbHelpers.addAndLoginAgent('Delivery Agent', _const.ACCESS_LEVEL.DeliveryAgent)
       agentObj.aid = agent.aid;
       agentObj.jar = agent.rpJar;
       await models()['WarehouseTest'].insertMany(warehouses);
@@ -356,35 +537,19 @@ describe('AgentFinishedDelivery in App', () => {
               discount_ref: 0
             },
             product_instance_id: mongoose.Types.ObjectId(products[0].instances[0]._id),
-            tickets: []
+            tickets: [{
+              is_processed: false,
+              status: _const.ORDER_LINE_STATUS.Delivered,
+              desc: null,
+              receiver_id: warehouses.find(x => x.is_hub)._id,
+              timestamp: new Date()
+            }]
           },
           tickets: [{
             is_processed: false,
-            status: _const.ORDER_STATUS.DeliverySet,
+            status: _const.ORDER_STATUS.Delivered,
             desc: null,
-            receiver_id: mongoose.Types.ObjectId(),
-            timestamp: new Date()
-          }]
-        }
-      });
-      await models()['OrderTest'].update({
-        _id: mongoose.Types.ObjectId(orders[1]._id),
-      }, {
-        $set: {
-          order_lines: {
-            product_id: mongoose.Types.ObjectId(products[0]._id),
-            campaign_info: {
-              _id: mongoose.Types.ObjectId(),
-              discount_ref: 0
-            },
-            product_instance_id: mongoose.Types.ObjectId(products[0].instances[0]._id),
-            tickets: []
-          },
-          tickets: [{
-            is_processed: false,
-            status: _const.ORDER_STATUS.DeliverySet,
-            desc: null,
-            receiver_id: mongoose.Types.ObjectId(),
+            receiver_id: agentObj._id,
             timestamp: new Date()
           }]
         }
@@ -393,13 +558,13 @@ describe('AgentFinishedDelivery in App', () => {
       deliveries = await models()['DeliveryTest'].insertMany([
         {
           to: {
+            warehouse_id: warehouses.find(x => x.is_hub)._id
+          },
+          from: {
             customer: {
               _id: orderData[0].customer_id,
               address: orderData[0].address
             }
-          },
-          from: {
-            warehouse_id: warehouses.find(x => x.is_hub)._id
           },
           order_details: [{
             order_line_ids: [
@@ -414,7 +579,7 @@ describe('AgentFinishedDelivery in App', () => {
             is_processed: false,
             _id: mongoose.Types.ObjectId(),
             status: _const.DELIVERY_STATUS.ended,
-            receiver_id: agentObj.aid,
+            receiver_id: warehouses.find(x => x.is_hub)._id,
             timestamp: new Date()
           }],
           delivery_agent: agentObj.aid,
@@ -426,7 +591,8 @@ describe('AgentFinishedDelivery in App', () => {
       done();
     } catch (err) {
       console.log(err);
-    };
+    }
+    ;
   }, 15000);
