--- spec/api/logistic/order/historyBox.spec.js
@@ -1,129 +1,0 @@
-const rp = require('request-promise');
-const lib = require('../../../../lib/index');
-const models = require('../../../../mongo/models.mongo');
-const mongoose = require('mongoose');
-const _const = require('../../../../lib/const.list');
-const warehouses = require('../../../../warehouses');
-const moment = require('moment');
-const utils = require('../utils');
-
-describe('POST Search OrderHistory', () => {
-
-  let salesManager = {
-    aid: null,
-    jar: null
-  };
-
-  let hubClerk = {
-    aid: null,
-    jar: null
-  };
-
-  let products, centralWarehouse, hubWarehouse;
-  beforeEach(async (done) => {
-    try {
-
-      await lib.dbHelpers.dropAll();
-
-      let warehouse = await models()['WarehouseTest'].insertMany(warehouses);
-      warehouse = JSON.parse(JSON.stringify(warehouse));
-
-      hubWarehouse = warehouse.find(x => x.is_hub && !x.has_customer_pickup);
-      centralWarehouse = warehouse.find(x => !x.is_hub && !x.has_customer_pickup);
-
-      let res1 = await lib.dbHelpers.addAndLoginAgent('salesManager', _const.ACCESS_LEVEL.SalesManager, centralWarehouse._id);
-      salesManager.aid = res1.aid;
-      salesManager.jar = res1.rpJar;
-
-      let res = await lib.dbHelpers.addAndLoginAgent('hclerk', _const.ACCESS_LEVEL.HubClerk, hubWarehouse._id);
-      hubClerk.aid = res.aid;
-      hubClerk.jar = res.rpJar;
-
-      products = await models()['ProductTest'].insertMany([
-        {
-          name: 'sample 1',
-          base_price: 30000,
-          article_no: "ar123",
-          instances: [
-            {
-              size: "9",
-              price: 2000,
-              barcode: '0394081341',
-            },
-            {
-              size: "10",
-              price: 2000,
-              barcode: '0394081342',
-            }
-          ]
-        }
-      ]);
-
-      done()
-    }
-    catch (err) {
-      console.log(err);
-    }
-  }, 15000);
-
-
-  it('only sales manager can see orders', async function (done) {
-    try {
-      this.done = done;
-
-      let orders = [];
-
-      orders.push({
-        customer_id: mongoose.Types.ObjectId(),
-        is_cart: false,
-        transaction_id: "xyz12213",
-        order_lines: [],
-        is_collect: true,
-        order_time: moment(),
-      });
-
-      for (let i = 0; i < 3; i++) { // add 3 order line of first product for order 1
-        orders[0].order_lines.push({
-          paid_price: 0,
-          product_id: products[0].id,
-          product_instance_id: products[0].instances[0].id,
-          adding_time: moment(),
-          cancel: false,
-          tickets: [
-            {
-              is_processed: false,
-              status: _const.ORDER_LINE_STATUS.DeliverySet,
-              desc: null,
-              receiver_id: hubWarehouse._id,
-              timestamp: moment()
-            }
-          ]
-        })
-      }
-
-      orders = await models()['OrderTest'].insertMany(orders);
-      orders = JSON.parse(JSON.stringify(orders));
-
-      let res = await rp({
-        method: 'post',
-        uri: lib.helpers.apiTestURL(`search/Ticket`),
-        body: {
-          options: {
-            type: 'OrdersHistory',
-          },
-          offset: 0,
-          limit: 10,
-        },
-        json: true,
-        // jar:
-        resolveWithFullResponse: true
-      });
-      this.fail('only sales manager can see history box');
-      done()
-    } catch (err) {
-      expect(err.statusCode).toBe(403);
-      done()
-    };
-  });
-
-});
